system-ificated the randomevents code so it's easy and safe to add new events
