+++ Assets/Scripts/RandomEvents.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f88a4f44a4418e48851378d959915cd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
