+++ Assets/Scripts/RandomEvents/Events.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b80cbecedfd3bf047b2db4d09621d98d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
