+++ Assets/Scripts/RandomEvents/Events/Negative.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 034f00abd75c17b46b9e5bc8261e4392
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
