+++ Assets/Scripts/RandomEvents/Events/Negative/MilitaryRequest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f704072e787702240bce41fc4011d3e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
