+++ Assets/Scripts/RandomEvents/Events/Negative/PirateAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8436bba06caf20144880dfdb8e4277be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
