+++ Assets/Scripts/RandomEvents/Events/Negative/SickCrewMember.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 10602e9bd885aea4585498cd9470268e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
