+++ Assets/Scripts/RandomEvents/Events/Negative/StormAtSea.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7a68408b4933d164f924781ee1382792
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
