+++ Assets/Scripts/RandomEvents/Events/Positive/AbandonedShip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1dde455cf98fe734984a9d74f0f189dd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
