+++ Assets/Scripts/RandomEvents/Events/Positive/FavorOfGods.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 21f122d03733dc344878f41885c2acb5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
