+++ Assets/Scripts/RandomEvents/Events/Positive/FriendlyShip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 456fbb10c13eee043ae66062044025ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
